Port amd64 kernel merge code into dev branch

+++ minix/drivers/storage/ahci/ahci.c
@@ -391,7 +391,7 @@ static int atapi_read_capacity(struct port_state *ps, int cmd)
 	dprintf(V_INFO,
 		("%s: medium detected (%u byte sectors, %llu MB size)\n",
 		ahci_portname(ps), ps->sector_size,
-		ps->lba_count * ps->sector_size / (1024*1024)));
+		(unsigned long long)(ps->lba_count * ps->sector_size / (1024*1024))));
 
 	return OK;
 }
@@ -1152,7 +1152,7 @@ static ssize_t port_transfer(struct port_state *ps, u64_t pos, u64_t eof,
 		return r;
 
 	dprintf(V_REQ, ("%s: %s for %lu bytes at pos %llx\n",
-		ahci_portname(ps), write ? "write" : "read", size, pos));
+		ahci_portname(ps), write ? "write" : "read", size, (unsigned long long)pos));
 
 	assert(ps->state == STATE_GOOD_DEV);
 	assert(ps->flags & FLAG_HAS_MEDIUM);
@@ -1425,7 +1425,7 @@ static void port_id_check(struct port_state *ps, int success)
 		if (ps->flags & FLAG_HAS_MEDIUM)
 			printf(", %u byte sectors, %llu MB size",
 				ps->sector_size,
-				ps->lba_count * ps->sector_size / (1024*1024));
+				(unsigned long long)(ps->lba_count * ps->sector_size / (1024*1024)));
 
 		printf("\n");
 	}
